Fast fertig
Überrarbeitet Run wird aus irgendwelchen gründen nicht ausgeführt

+++ source/Natural Selection Sim/Natural Selection Sim/Logic/Herbivore.cs
@@ -2,7 +2,7 @@ namespace Natural_Selection_Sim
 {
     public class Herbivore : Entity
     {
-        public Herbivore(float b, float d, float m, float s, float z)
+        public Herbivore(double b, double d, double m, double s, double z)
             : base(b, d, m, s, z) { }
 
         public Herbivore(Entity parent)
